pyproject.toml
[tool.poetry]
name = "marine-flow"
version = "0.1.0"
description = ""
authors = ["Steffen Henkelmann <coder.henkelmann@gmail.com>"]
readme = "README.md"
packages = [{include = "marine_flow", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
greenlet = "*"
google-cloud-firestore = "^2.7.0"
torch = "^1.12.1"

[tool.poetry.group.dev.dependencies]
darglint = "^1.8.1"
flake8 = "^5.0.4"
black = "^22.8.0"
mypi = "^0.0.3"
pylint = "^2.15.2"
pydocstyle = "^6.1.1"
safety = "^2.1.1"
isort = "^5.10.1"
mypy = "^0.971"
pytest = "^7.1.3"
pytest-cov = "^3.0.0"


[tool.poetry.scripts]
marine-etl-flow = "flows.pipeline_etl:etl_flow"
marine-ml-flow = "flows.pipeline_ml:ml_training_flow"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
# https://github.com/psf/black
line-length = 100
target-version = ["py310"]

[tool.isort]
# https://github.com/timothycrosley/isort/
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
indent = 4
force_grid_wrap = 0
use_parentheses = true
line_length = 100
